Use utf-8 charset
This sets the charset in editorconfig to utf-8. Previously, there was a mixture of file encodings with some including a BOM (byte order mark) and some not. Most tools will not show the BOM, and so the diff of this commit may appear mostly empty. All the files in the diff, apart from .editorconfig, simply had the BOM removed.