Always include displayname in external providers
There's an edge case where we don't set the displayname, which normally doesn't matter because our templates would immediately challenge instead of needing to display the external provider. However, some users have run into situations where it caused errors for them.

This is a recreation of a community PR to fix that issue. I've rebased onto the latest in main, and applied the fix to all hosts and templates.

Thanks to @stefannikolei for the original contribution!